+++ CHANGELOG.md
@@ -6,5 +6,12 @@ All notable changes to this project will be documented in this file.
 
 ### Initial Release
 
+## 0.0.1-beta.9 (2020-09-01)
+* Added remote control layers
+* Improved config schema
+* Improved ping code
+* Bumped dependencies
+* Updated Readme
+
 ## 0.0.1-beta.x (2020-04-21)
 * Alpha release for experimentation